Refactor namespaces and organize project files
- Moved `AddUpdateBook` to `BookstoreApp.Forms` and `Book` to `BookstoreApp.Models` for better organization.
- Updated constructors and methods in `AddUpdateBook.cs` to reflect new namespaces and improved form title handling.
- Enhanced validation for ISBN and price fields in the save button click event.
- Added a description property to the `Book` class and retained documentation.
- Updated `BookDb.cs` and `BookStoreDb.cs` to use the new `BookstoreApp.Models` namespace.
- Adjusted `Main.cs` to include event handlers for book management and maintain asynchronous loading of books.
- Updated designer files and resource files to ensure proper association with refactored classes.
- Overall improvements enhance code maintainability and readability.
