+++ BookstoreApp/Forms/AddUpdateBook.resx
