+++ BookstoreApp/Forms/Main.cs
@@ -1,4 +1,6 @@
 using BookstoreApp.Database;
+using BookstoreApp.Forms;
+using BookstoreApp.Models;
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
